+++ opal/corelib/basic_object.rb
@@ -52,7 +52,22 @@ class BasicObject
           result;
 
       block.$$s = null;
-      result = block.call(self, self);
+      
+      // need to pass $$eval so that method definitions know if this is being done on a class/module. Cannot be compiler driven since
+      // send(:instance_eval) needs to work
+      if (self.$$is_class || self.$$is_module) {
+        self.$$eval = true;
+        try {          
+          result = block.call(self, self);
+        }
+        finally {
+          self.$$eval = false;
+        }
+      }
+      else {
+        result = block.call(self, self);
+      }
+      
       block.$$s = old;
 
       return result;
@@ -65,9 +80,22 @@ class BasicObject
     %x{
       var block_self = block.$$s,
           result;
-
+          
       block.$$s = null;
-      result = block.apply(self, args);
+      
+      if (self.$$is_class || self.$$is_module) {
+        self.$$eval = true;
+        try {        
+          result = block.apply(self, args);
+        }
+        finally {
+          self.$$eval = false;
+        }        
+      }
+      else {
+        result = block.apply(self, args);
+      }
+
       block.$$s = block_self;
 
       return result;