+++ spec/opal/core/exception_spec.rb
@@ -11,3 +11,10 @@ describe "Exception" do
     ExceptionSubclassTest.new.custom_method.should == 42
   end
 end
+
+describe "Native exception" do
+  it "handles messages for native exceptions" do
+    exception = `new Error("native message")`
+    exception.message.should == "native message"
+  end
+end